Tests/LibWeb/Text/expected/svg-abspos.txt

x: 260
y: 0
width: 540
height: 390
x: 0
y: 210
width: 540
height: 390
